Splits testLargeCase in metric_ops_test into a dedicated file for slow-running tests and re-enables it as a 'large' test.
PiperOrigin-RevId: 200440883
